% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/detectOutliers.R
\name{outliers}
\alias{outliers}
\title{Identifies outliers for plot_interactive.fosr()}
\usage{
outliers(data, factor = 1.5)
}
\arguments{
\item{name}{variable name}

\item{variable}{variable values from dataset}
}
\description{
Internal method that assigns band depth values to curves based on exact fast MBD computation (Sun & Genton, 2012).
A dataframe of residuals is passed as an argument, and depths and outlying curves are returned
}
\author{
Julia Wrobel \email{jw3134@cumc.columbia.edu}
}

