% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_interactive.fosr.R
\name{plot_interactive.fosr}
\alias{plot_interactive.fosr}
\title{Interactive Plotting for Functional-on-Scalar Regressions}
\usage{
\method{plot_interactive}{fosr}(fosr.obj, xlab = "", ylab = "",
  title = "")
}
\arguments{
\item{fosr.obj}{fosr object to be plotted.}

\item{xlab}{x axis label}

\item{ylab}{y axis label}

\item{title}{plot title}
}
\description{
Produces an interactive plot illustrating a function-on-scalar
regression analysis. So far, there's very little here.
}
\examples{
library(refund)
library(dplyr)

set.seed(101)

data(DTI)
DTI = subset(DTI, select = c(cca, case, pasat))
DTI = DTI[complete.cases(DTI),]
DTI$gender = sample(c(0,1), dim(DTI)[1], replace = TRUE)

fosr.dti = fosr_gls(cca ~ pasat + gender, data = DTI)

plot_interactive(fosr.dti)
}
\author{
Julia Wrobel \email{jw3134@cumc.columbia.edu},
Jeff Goldsmith \email{jeff.goldsmith@columbia.edu}
}
\seealso{
\code{\link{plot.interactive}}
}

