% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_interactive.fpca.R
\name{plot_interactive.fpca}
\alias{plot_interactive.fpca}
\title{Interactive Plotting for Functional Principal Component Analysis}
\usage{
\method{plot_interactive}{fpca}(fpca.obj, xlab = "", ylab = "",
  title = "")
}
\arguments{
\item{fpca.obj}{fpca object to be plotted.}

\item{xlab}{x axis label}

\item{ylab}{y axis label}

\item{title}{plot title}
}
\description{
Produces an interactive plot illustrating a functional principal component
analysis.
}
\examples{
library(refund)
data(cd4)
fpca.cd4 = fpca(cd4, var=TRUE)
plot_interactive(fpca.cd4)
}
\author{
Julia Wrobel \email{jw3134@cumc.columbia.edu},
Jeff Goldsmith \email{jeff.goldsmith@columbia.edu}
}
\seealso{
\code{\link{plot.interactive}}
}

