% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_interactive.R
\name{plot_interactive}
\alias{plot_interactive}
\title{Interactive Plotting for Functional Data}
\usage{
plot_interactive(x, ...)
}
\arguments{
\item{x}{object to be plotted. Currenlty, only allowed data type is "fpca".}

\item{...}{additional arguments passed to plotting functions}
}
\description{
Generic function for interactive plotting of functional data analysis results
}
\examples{
library(refund)

data(cd4)
fpca.cd4 = fpca.sc(cd4, var=TRUE)
class(fpca.cd4) = "fpca"
plot_interactive(fpca.cd4)

data(DTI)
fpca.dti = fpca.sc(DTI$cca)
class(fpca.dti) = "fpca"
plot_interactive(fpca.dti)
}
\author{
Jeff Goldsmith \email{jeff.goldsmith@columbia.edu},
Julia Wrobel \email{jw3134@cumc.columbia.edu}
}
\seealso{
\code{\link{plot.interactive.fpca}}
}

