% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_interactive.fosr.R
\name{plot_interactive.fosr}
\alias{plot_interactive.fosr}
\title{Interactive Plotting for Functional-on-Scalar Regressions}
\usage{
\method{plot_interactive}{fosr}(fosr.obj, xlab = "", ylab = "",
  title = "")
}
\arguments{
\item{fosr.obj}{fosr object to be plotted.}

\item{xlab}{x axis label}

\item{ylab}{y axis label}

\item{title}{plot title}
}
\description{
Produces an interactive plot illustrating a function-on-scalar
regression analysis.
}
\author{
Jeff Goldsmith \email{jeff.goldsmith@columbia.edu},
Julia Wrobel \email{jw3134@cumc.columbia.edu}
}
\seealso{
\code{\link{plot_interactive}}
}

