% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_interactive.R
\name{plot_interactive}
\alias{plot_interactive}
\title{Interactive Plotting for Functional Data}
\usage{
plot_interactive(x, ...)
}
\arguments{
\item{x}{object to be plotted. Currenlty, allowed data types are \code{fpca} and \code{fosr}.}

\item{...}{additional arguments passed to plotting functions}
}
\description{
Generic function for interactive plotting of functional data analysis results
}
\examples{
library(refund)
library(dplyr)


##### FPCA Example #####

data(cd4)
fpca.cd4 = fpca(cd4, var=TRUE)
plot_interactive(fpca.cd4)


##### FoSR Example #####

data(DTI)
DTI = subset(DTI, select = c(cca, case, pasat))
DTI = DTI[complete.cases(DTI),]
DTI$gender = factor(sample(c("male","female"), dim(DTI)[1], replace = TRUE))
DTI$status = factor(sample(c("RRMS", "SPMS", "PPMS"), dim(DTI)[1], replace = TRUE))

fosr.dti1 = fosr_gls(cca ~ pasat, data = DTI)
plot_interactive(fosr.dti1)

fosr.dti2 = fosr_gls(cca ~ pasat * gender + status, data = DTI)
plot_interactive(fosr.dti2)


##### FoSR Example with outliers #####

DTI$cca[1,] = DTI$cca[1,] + .3
DTI$cca[2,] = DTI$cca[2,] + .3

fosr.dti3 = fosr_gls(cca ~ pasat * gender + status, data = DTI)
plot_interactive(fosr.dti3)
}
\author{
Jeff Goldsmith \email{jeff.goldsmith@columbia.edu},
Julia Wrobel \email{jw3134@cumc.columbia.edu}
}
\seealso{
\code{\link{plot.interactive.fpca}}
}

